+++ proguardprocessor/factory.go
@@ -31,7 +31,7 @@ func createDefaultConfig() component.Config {
 		MethodsAttributeKey:             "exception.structured_stacktrace.methods",
 		LinesAttributeKey:               "exception.structured_stacktrace.lines",
 		SourceFilesAttributeKey:         "exception.structured_stacktrace.source_files",
-		OutputStackTraceKey:             "exception.stacktrace",
+		StackTraceAttributeKey:          "exception.stacktrace",
 		ExceptionTypeAttributeKey:       "exception.type",
 		ExceptionMessageAttributeKey:    "exception.message",
 		PreserveStackTrace:              true,